refactor additionalSkip
